Test Predict

+++ modules/tfrecordhandler_DP.py
@@ -244,7 +244,7 @@ class TFRReader():
         appearance_image = tf.io.decode_jpeg(parsed_ex['Appearance'], channels=3)
         appearance_image = tf.reshape(appearance_image, shape=(im_height, im_width, im_depth))
         
-        label = tf.cast(parsed_ex['Labels'], dtype = tf.int32)
+        label = tf.cast(parsed_ex['Labels'], dtype = tf.float32)
         
         return (motion_image, appearance_image), label, im_name, frames
